+++ Package.swift
@@ -12,7 +12,7 @@ let package = Package(
     dependencies: [
         .package(url: "https://github.com/groue/GRDB.swift", from: "6.0.0"),
         .package(url: "https://github.com/apple/swift-argument-parser", from: "1.3.0"),
-        .package(url: "https://github.com/P-H-C/phc-winner-argon2.git", branch: "master"),
+        .package(url: "https://github.com/P-H-C/phc-winner-argon2.git", revision: "f57e61e19229e23c4445b85494dbf7c07de721cb"),
         .package(url: "https://github.com/xnth97/SymbolPicker.git", .upToNextMajor(from: "1.6.0")),
     ],
     targets: [